displacement and renaming of tool classes
