+++ src/main/java/tools/Evaluation.java
@@ -1,4 +1,4 @@
-package algorithm;
+package tools;
 
 import algorithm.NEAT.Individual;
