+++ src/test/java/algorithm/running_choice/TestRunningAlgorithm.java
@@ -4,7 +4,6 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
-import algorithm.Evaluation;
 import algorithm.LearningAlgorithm;
 import algorithm.NEAT.Individual;
 import algorithm.NEAT.NEATAlgorithm;
@@ -12,6 +11,7 @@ import algorithm.NEAT.reproduction.Elitism;
 import algorithm.NEAT.reproduction.ReproductionAlgorithm;
 import brain.Brain;
 import brain.LayeredBrain;
+import tools.Evaluation;
 
 class TestRunningAlgorithm {
